UserService/pyproject.toml
[tool.poetry]
name = "userservice"
version = "0.1.0"
description = ""
authors = ["muneeb <muneebxhasan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.2.2"
fastapi = "^0.111.1"
sqlmodel = "^0.0.19"
aiokafka = "^0.11.0"
psycopg = "^3.2.1"
passlib = "^1.7.4"
pycryptodome = "^3.20.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
